Moved query conversion stuff into its own file
